[issue-4] 리덕스 잠시중단
